+++ lib/tiled/builder/tiled_world_builder.dart
@@ -33,13 +33,6 @@ typedef ObjectBuilder = GameComponent Function(
 class TiledWorldBuilder {
   static const ABOVE_TYPE = 'above';
   static const DYNAMIC_ABOVE_TYPE = 'dynamicAbove';
-  static const GIT_ROTATE_180 = 3221225472;
-  static const GIT_ROTATE_90 = 2684354560;
-  static const GIT_ROTATE_270 = 1610612736;
-  static const GIT_FLIP_HORIZONTAL = 2147483648;
-  static const GIT_FLIP_VERTICAL = 1073741824;
-  static const GIT_FLIP_HORIZONTAL_270 = 536870912;
-  static const GIT_FLIP_HORIZONTAL_90 = 3758096384;
 
   final Vector2? forceTileSize;
   final ValueChanged<Object>? onError;
@@ -258,36 +251,8 @@ class TiledWorldBuilder {
   }
 
   TiledItemTileSet? _getDataTile(int gid) {
-    int index = 0;
-    double angle = 0;
-    bool isFlipX = false;
-    bool isFlipY = false;
-    if (gid > GIT_FLIP_HORIZONTAL_90) {
-      isFlipX = true;
-      angle = 1.5708;
-      index = gid - GIT_FLIP_HORIZONTAL_90;
-    } else if (gid > GIT_ROTATE_180) {
-      angle = 3.14159;
-      index = gid - GIT_ROTATE_180;
-    } else if (gid > GIT_ROTATE_90) {
-      angle = 1.5708;
-      index = gid - GIT_ROTATE_90;
-    } else if (gid > GIT_FLIP_HORIZONTAL) {
-      isFlipX = true;
-      index = gid - GIT_FLIP_HORIZONTAL;
-    } else if (gid > GIT_ROTATE_270) {
-      angle = 4.71239;
-      index = gid - GIT_ROTATE_270;
-    } else if (gid > GIT_FLIP_VERTICAL) {
-      isFlipY = true;
-      index = gid - GIT_FLIP_VERTICAL;
-    } else if (gid > GIT_FLIP_HORIZONTAL_270) {
-      isFlipX = true;
-      angle = 4.71239;
-      index = gid - GIT_FLIP_HORIZONTAL_270;
-    } else {
-      index = gid;
-    }
+    final gidInfo = TileLayer.getGidInfo(gid);
+    int index = gidInfo.index;
 
     TileSetDetail? tileSetContain;
     String pathTileset = '';
@@ -374,9 +339,9 @@ class TiledWorldBuilder {
         properties: object.properties,
         sprite: sprite,
         animation: animation,
-        angle: angle,
-        isFlipHorizontal: isFlipX,
-        isFlipVertical: isFlipY,
+        angle: gidInfo.angle,
+        isFlipHorizontal: gidInfo.isFlipX,
+        isFlipVertical: gidInfo.isFlipY,
       );
     } else {
       return null;